7601 Added log file generation changes  AB#7601 AB#8255

+++ UKHO.ExchangeSetService.API/UKHO.ExchangeSetService.CleanUpJob/Helpers/AzureDeleteFileSystemHelper.cs
@@ -45,16 +45,20 @@ namespace UKHO.ExchangeSetService.CleanUpJob.Helpers
                             {
                                 var batchidFile = dir.Name;
                                 string batchId = new DirectoryInfo(batchidFile).Name + ".json";
-
-                                await azureBlobStorageClient.DeleteContainerFile(storageAccountConnectionString, containerName, batchId);                               
-                                dir.Delete(true);
-                                
-                                logger.LogInformation(EventIds.DeleteHistoricContainerFile.ToEventId(), "SCS response from the container for BatchId:{BatchId} deleted successfully.", batchId);
+                                var response =await azureBlobStorageClient.DeleteContainerFile(storageAccountConnectionString, containerName, batchId);
+                                if (response)
+                                {
+                                    dir.Delete(true);
+                                    logger.LogInformation(EventIds.DeleteHistoricContainerFile.ToEventId(), "SCS response from the container file deleted successfully for BatchId:{BatchId}.", batchId);
+                                }
+                                else
+                                {
+                                    logger.LogInformation(EventIds.DeleteHistoricContainerFileNotFound.ToEventId(), "SCS response from the container file not found for BatchId:{BatchId}.", batchId);
+                                }
                             }
                             di.Delete(true);
                             deleteStatus = true;
-                            
-                            logger.LogInformation(EventIds.DeleteHistoricFolder.ToEventId(), "Historic folder for DateFolder:{dateFolder} deleted successfully.", dateFolder);
+                            logger.LogInformation(EventIds.DeleteHistoricFolder.ToEventId(), "Historic folder deleted successfully for DateFolder:{dateFolder}.", dateFolder);
                         }
                     }
                 }
