+++ UKHO.ExchangeSetService.API/UKHO.ExchangeSetService.Common/Helpers/AzureBlobStorageClient.cs
@@ -32,13 +32,22 @@ namespace UKHO.ExchangeSetService.Common.Helpers
         {
              return await cloudBlockBlob.DownloadTextAsync();
         }
-        public async Task DeleteContainerFile(string storageAccountConnectionString, string containerName, string batchId)
+        public async Task<bool> DeleteContainerFile(string storageAccountConnectionString, string containerName, string batchId)
         {
             CloudStorageAccount cloudStorageAccount = CloudStorageAccount.Parse(storageAccountConnectionString);
             CloudBlobClient _blobClient = cloudStorageAccount.CreateCloudBlobClient();
             CloudBlobContainer _cloudBlobContainer = _blobClient.GetContainerReference(containerName);
             CloudBlockBlob _blockBlob = _cloudBlobContainer.GetBlockBlobReference(batchId);
-            await _blockBlob.DeleteIfExistsAsync();
+            bool checkBlobReference = await _blockBlob.ExistsAsync();
+            if (checkBlobReference)
+            {
+                await _blockBlob.DeleteIfExistsAsync();
+                return true;
+            }
+            else
+            {
+                return false;
+            }
         }
     }
 }
